[chore] consolidate input-stanza to receiver mapping into a shared builder
The conf-file readers and scheme dispatch were duplicated in
internal/collector and pkg/splunkinputsreceiver. They now live in a single
internal/tabuilder package used by both. No behavior change.
