+++ internal/collector/collector_test.go
@@ -156,45 +156,6 @@ func TestRunScriptedInputs(t *testing.T) {
 	}, 2*time.Second, 10*time.Millisecond)
 }
 
-func TestReadTransforms(t *testing.T) {
-	rootDir := filepath.Join("testdata", "transforms")
-	tests := []struct {
-		name          string
-		path          string
-		expectedName  string
-		expectedRegex string
-	}{
-		{
-			name:          "default",
-			path:          filepath.Join(rootDir, "default"),
-			expectedName:  "example_default",
-			expectedRegex: "default",
-		},
-		{
-			name:          "local",
-			path:          filepath.Join(rootDir, "local"),
-			expectedName:  "example_local",
-			expectedRegex: "local",
-		},
-		{
-			name:          "both",
-			path:          filepath.Join(rootDir, "both"),
-			expectedName:  "example_local2",
-			expectedRegex: "local",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			transforms, err := readTransforms(test.path)
-			require.NoError(t, err)
-			require.Len(t, transforms, 1)
-			require.Equal(t, test.expectedName, transforms[0].Name)
-			require.Equal(t, test.expectedRegex, transforms[0].Regex)
-		})
-	}
-}
-
 func TestUseTCP(t *testing.T) {
 	rootDir := filepath.Join("testdata", "tcp")
 	logsSink := &consumertest.LogsSink{}
