+++ pkg/splunkinputsreceiver/go.mod
@@ -69,3 +69,5 @@ require (
 	google.golang.org/protobuf v1.36.11 // indirect
 	gopkg.in/ini.v1 v1.67.3 // indirect
 )
+
+replace github.com/splunk/tarunner => ../../
